# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API do Duora. Duas portas de entrada com as mesmas regras de rota: cookie de sessão do front web (BFF), com o token CSRF do cookie XSRF-TOKEN no header X-XSRF-TOKEN em toda mutação, e Authorization: Bearer com o access token do Entra External ID. Erros em application/problem+json.",
    "title" : "Duora API",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "/",
    "x-internal" : false
  } ],
  "security" : [ {
    "bearer" : [ ]
  }, {
    "session" : [ ]
  } ],
...
  "paths" : {
...
    "/api/events/{eventId}/registration" : {
      "delete" : {
        "description" : "Idempotente: sem inscrição, também responde 204. Só até o evento começar.",
        "operationId" : "cancelMyRegistration",
        "parameters" : [ {
          "description" : "Id do evento",
          "in" : "path",
          "name" : "eventId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "maxLength" : 36,
            "minLength" : 36
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Quem chama não está inscrito no evento",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Id que não é UUID",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Sem credencial válida",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Sessão web sem o token CSRF no header X-XSRF-TOKEN",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Evento inexistente ou rascunho",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "O evento já começou",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Erro inesperado",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          }
        },
        "summary" : "Cancela a própria inscrição no evento",
        "tags" : [ "events" ]
      },
      "get" : {
        "description" : "Só a de quem chama: não há como ler a inscrição de outra pessoa. Sem inscrição, 404, exista o evento ou não.",
        "operationId" : "getMyRegistration",
        "parameters" : [ {
          "description" : "Id do evento",
          "in" : "path",
          "name" : "eventId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "maxLength" : 36,
            "minLength" : 36
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description" : "A inscrição",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Id que não é UUID",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Sem credencial válida",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Quem chama não está inscrito nesse evento",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Erro inesperado",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          }
        },
        "summary" : "Lê a própria inscrição no evento",
        "tags" : [ "events" ]
      },
      "put" : {
        "description" : "Sem corpo. Idempotente pela chave evento + conta: repetir devolve a mesma inscrição, com a mesma data, por isso dispensa If-Match e Idempotency-Key (docs/adr/0016). Exige perfil completo e 18 anos.",
        "operationId" : "registerForEvent",
        "parameters" : [ {
          "description" : "Id do evento",
          "in" : "path",
          "name" : "eventId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid",
            "maxLength" : 36,
            "minLength" : 36
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description" : "A inscrição que já existia",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description" : "A inscrição criada",
            "headers" : {
              "Location" : {
                "description" : "Endereço da inscrição",
                "required" : true,
                "schema" : {
                  "type" : "string",
                  "format" : "uri",
                  "maxLength" : 2048
                },
                "style" : "simple"
              },
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Id que não é UUID",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Sem credencial válida",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Perfil incompleto ou menor de 18 anos, ou sessão web sem o token CSRF no header X-XSRF-TOKEN",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Evento inexistente ou rascunho",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Evento cancelado, já começado ou lotado",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Erro inesperado",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "O evento está ocupado com outras inscrições; nada foi gravado",
            "headers" : {
              "Retry-After" : {
                "description" : "Segundos até tentar de novo",
                "required" : true,
                "schema" : {
                  "type" : "integer",
                  "format" : "int32",
                  "maximum" : 1,
                  "minimum" : 1
                },
                "style" : "simple"
              },
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          }
        },
        "summary" : "Inscreve quem chama no evento",
        "tags" : [ "events" ]
      }
    },
...
    "/api/me/registrations" : {
      "get" : {
        "description" : "Em eventos que ainda não acabaram, inclusive cancelados e em andamento, do início mais próximo ao mais distante, paginadas por cursor. A última página vem com nextPageToken null.",
        "operationId" : "listMyRegistrations",
        "parameters" : [ {
          "description" : "Quantas inscrições no máximo nesta página",
          "in" : "query",
          "name" : "maxPageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "maximum" : 50,
            "minimum" : 1
          }
        }, {
          "description" : "O nextPageToken da página anterior; ausente na primeira",
          "in" : "query",
          "name" : "pageToken",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 90,
            "pattern" : "^[A-Za-z0-9_-]+$"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseMyRegistrationResponse"
                }
              }
            },
            "description" : "Uma página das inscrições",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "maxPageSize fora de 1 a 50, ou pageToken que a API não gerou",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Sem credencial válida",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Erro inesperado",
            "headers" : {
              "X-Request-Id" : {
                "$ref" : "#/components/headers/RequestId"
              }
            }
          }
        },
        "summary" : "Lista as próprias inscrições",
        "tags" : [ "events" ]
      }
    },
...
  },
  "components" : {
    "headers" : {
      "RequestId" : {
        "description" : "Correlation ID da requisição (trace id W3C); informe-o ao reportar um erro",
        "required" : true,
        "schema" : {
          "type" : "string",
          "maxLength" : 32,
          "minLength" : 32,
          "pattern" : "^[0-9a-f]{32}$"
        }
      }
    },
    "schemas" : {
...
      "MyRegistrationResponse" : {
        "type" : "object",
        "properties" : {
          "endsAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Fim do evento, em UTC",
            "maxLength" : 35
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Id do evento",
            "maxLength" : 36,
            "minLength" : 36
          },
          "eventStatus" : {
            "type" : "string",
            "description" : "Estado do evento: um cancelado continua na lista, para a pessoa saber",
            "enum" : [ "PUBLISHED", "CANCELLED" ]
          },
          "registeredAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Quando a inscrição foi feita",
            "maxLength" : 35
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Início do evento, em UTC",
            "maxLength" : 35
          },
          "title" : {
            "type" : "string",
            "description" : "Título do evento",
            "maxLength" : 80
          }
        },
        "required" : [ "endsAt", "eventId", "eventStatus", "registeredAt", "startsAt", "title" ]
      },
...
      "PageResponseMyRegistrationResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "description" : "Itens da página, pelo início do evento",
            "items" : {
              "$ref" : "#/components/schemas/MyRegistrationResponse"
            },
            "maxItems" : 50
          },
          "nextPageToken" : {
            "type" : [ "string", "null" ],
            "description" : "Token da próxima página, ou null na última",
            "maxLength" : 90,
            "pattern" : "^[A-Za-z0-9_-]+$"
          }
        },
        "required" : [ "items", "nextPageToken" ]
      },
      "ProblemDetail" : {
        "type" : "object",
        "description" : "Erro no formato RFC 9457. Sem stack trace, SQL nem nome de classe.",
        "properties" : {
          "detail" : {
            "type" : "string",
            "maxLength" : 1000
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference",
            "maxLength" : 1000
          },
          "requestId" : {
            "type" : "string",
            "description" : "Correlation ID, o mesmo do header X-Request-Id; vem nos erros inesperados (500)",
            "maxLength" : 32,
            "minLength" : 32,
            "pattern" : "^[0-9a-f]{32}$"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 599,
            "minimum" : 100
          },
          "title" : {
            "type" : "string",
            "maxLength" : 1000
          },
          "type" : {
            "type" : "string",
            "format" : "uri-reference",
            "maxLength" : 1000
          }
        },
        "required" : [ "status", "title" ]
      },
...
      "RegistrationResponse" : {
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Id do evento",
            "maxLength" : 36,
            "minLength" : 36
          },
          "registeredAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Quando a inscrição foi feita; não muda nas repetições do PUT",
            "maxLength" : 35
          }
        },
        "required" : [ "eventId", "registeredAt" ]
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "Access token da duora-api emitido pelo Entra External ID, validado conforme a RFC8725: algoritmo fixo (RS256) e iss, aud, exp e oid obrigatórios.",
        "scheme" : "bearer",
        "type" : "http"
      },
      "session" : {
        "description" : "Sessão do front web, criada pelo login em /oauth2/authorization/entra",
        "in" : "cookie",
        "name" : "__Host-DUORA_SESSION",
        "type" : "apiKey"
      }
    }
  }
}